Factor out editable-info.tsx.
